chore: split PrivacyTab into PrivacyAutoLock, PrivacyBiometric, PrivacyTwoFactor, PrivacyDataManagement, PrivacyTransparency
660-line PrivacyTab.tsx extracted into 5 focused files; orchestrator now ~130 lines. Biometric state/effect encapsulated in PrivacyBiometric, reset confirm state in PrivacyDataManagement, 2FA modals remain in PrivacyTab alongside use2FASetup hook state.
